fix: arena cap uint32_t overflow causing OOM loop + mmap vectors
storage_hart: arena_cap/arena_used were uint32_t. When a whale
contract's arena grew past 2GB, pool_alloc rounded to 4GB
(size class 32), but (uint32_t)4GB = 0. This caused an infinite
grow loop — each cycle wasted 4GB from the pool until OOM.

Fix: widen arena_cap and arena_used to uint64_t. Switch arena
growth to linear +256MB steps above 1GB (avoids copying gigabytes
for a few bytes of new storage).

state: switch account and resource vectors from realloc (copies
everything, 2x temp spike) to mmap/mremap (zero copy, remaps
page tables). Growth factor changed from 2x to 1.5x.

LTO parallelism reduced from auto (7) to 2 to prevent OOM during
compilation while chain_replay is running.

+++ database/include/storage_hart.h
@@ -45,8 +45,8 @@ void storage_hart_pool_sync(storage_hart_pool_t *pool);
 
 typedef struct {
     uint64_t arena_offset;  /* byte offset of arena in pool file (0 = none) */
-    uint32_t arena_used;    /* bytes used in arena */
-    uint32_t arena_cap;     /* allocated capacity (bytes) */
+    uint64_t arena_used;    /* bytes used in arena */
+    uint64_t arena_cap;     /* allocated capacity (bytes) */
     uint32_t count;         /* number of live entries (leaves) */
     uint32_t root_ref;      /* ref to root node within arena (0 = empty) */
     /* Per-type freelist heads (arena-relative offsets, 0 = empty) */
